draw, r300, util: safety gates for NIR exec, blitter, and admission
Live-main review findings left host OOB and fail-open paths in the new
direct-NIR vertex executor, blitter NIR twins, and r300 diagnostic/admission
helpers.  Bound SSA/I/O/UBO access to the TGSI-compatible zero-fill
contract, stage loop-header phis as parallel copies, and seed load_draw_id
from draw->pt.user.drawid.  Gate blitter NIR on VS+FS NIR support, lower
samplers before gather_info, optimize finalize_nir, keep PIPE_BUFFER on TGSI,
and mark binding 0 used.  r300: exact R300_SWTCL_DUMMY_TEXCOORD=1, unbind
clears FS/VS constants, destroy drops fs_const0_app, measure_nir_admission
rejects NULL, multipass cut caps max_cands, inspect uses IT_COUNT macros,
front-color/FACE verdict hygiene, LAST_VEC bounds, MALLOC/FREE, and
compute offset equality includes offset_shift plus lowered system-value roots.

+++ src/gallium/winsys/radeon/drm/radeon_drm_bo.c
@@ -8,6 +8,7 @@
 
 #include "util/u_hash_table.h"
 #include "util/u_memory.h"
+#include "util/u_debug.h"
 #include "util/u_thread.h"
 #include "util/os_mman.h"
 #include "util/os_time.h"
@@ -87,7 +88,8 @@ static void radeon_real_bo_wait_idle(struct radeon_bo *bo)
    if (debug_get_bool_option("R300_TRACE_WAIT_IDLE", false)) {
       fprintf(stderr,
               "R300_WAIT_IDLE handle=%u size=%"PRIu64" initial_domain=%u va=0x%"PRIx64"\n",
-              bo->handle, (uint64_t)bo->base.size, bo->initial_domain, bo->va);
+              bo->handle, (uint64_t)bo->base.size, (unsigned)bo->initial_domain,
+              bo->va);
       fflush(stderr);
    }
